管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -42,6 +42,9 @@ public class EmployeeAction extends ActionBase {
      */
     public void index() throws ServletException, IOException {
 
+        //管理者かどうかのチェック //追記
+        if (checkAdmin()) { //追記
+
         //指定されたページ数の一覧画面に表示するデータを取得
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
@@ -64,6 +67,8 @@ public class EmployeeAction extends ActionBase {
         //一覧画面を表示
         forward(ForwardConst.FW_EMP_INDEX);
 
+        } //追記
+
     }
 
 
@@ -74,11 +79,15 @@ public class EmployeeAction extends ActionBase {
      */
     public void entryNew() throws ServletException, IOException {
 
+        // 管理者かどうかのチェック //追記
+        if (checkAdmin()) { //追記
         putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
         putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
 
         //新規登録画面を表示
         forward(ForwardConst.FW_EMP_NEW);
+
+        }//追記
     }
 
     /**
@@ -89,7 +98,7 @@ public class EmployeeAction extends ActionBase {
     public void create() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) {  //追記
 
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
@@ -138,6 +147,9 @@ public class EmployeeAction extends ActionBase {
      */
     public void show() throws ServletException, IOException {
 
+        //管理者かどうかのチェック //追記
+        if (checkAdmin()) { //追記
+
         //idを条件に従業員データを取得する
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -152,6 +164,8 @@ public class EmployeeAction extends ActionBase {
 
         //詳細画面を表示
         forward(ForwardConst.FW_EMP_SHOW);
+
+        } //追記
     }
 
     /**
@@ -160,6 +174,9 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void edit() throws ServletException, IOException {
+
+        //管理者かどうかのチェック //追記
+        if (checkAdmin()) { //追記
 
         //idを条件に従業員データを取得する
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -177,6 +194,8 @@ public class EmployeeAction extends ActionBase {
         //編集画面を表示する
         forward(ForwardConst.FW_EMP_EDIT);
 
+        } //追記
+
     }
 
     /**
@@ -187,7 +206,7 @@ public class EmployeeAction extends ActionBase {
     public void update() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) { //追記
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -233,7 +252,7 @@ public class EmployeeAction extends ActionBase {
     public void destroy() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) { //追記
 
             //idを条件に従業員データを論理削除する
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -245,6 +264,31 @@ public class EmployeeAction extends ActionBase {
             redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
         }
     }
+
+    /**
+     * ログイン中の従業員が管理者かどうかチェックし、管理者でなければエラー画面を表示
+     * true: 管理者 false: 管理者ではない
+     * @throws ServletException
+     * @throws IOException
+     */
+    private boolean checkAdmin() throws ServletException, IOException {
+
+        //セッションからログイン中の従業員情報を取得
+        EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+        //管理者でなければエラー画面を表示
+        if (ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+
+        } else {
+
+            return true;
+        }
+
+    }
+
 
 
 }
